requirements.txt
Flask==3.0.0
pymongo==4.6.0
python-dotenv==1.0.1
Werkzeug==3.0.1
WTForms==3.1.1
requests==2.31.0
Pillow==10.3.0
gunicorn==21.2.0